Adding New Form Organization
The new way of organizing forms allows for multiple companies and facilities to use the same form as the old way made every facility use the same exact forms to keep track of what was submitted. Started to edit the Client Forms as well. Got to a good stopping point.