Facility Forms, Facility List Fixed, Form Progress
All these pages are working. Need to add the function that allows for the Progress page to update when a form is added that already has forms submitted (sounds weird but it could happen by accident if they accidentally remove the form from the list. Right now if you updated the forms list, it deletes every entry and then resaves the list. Maybe can find a way to fix that. Still need to remove certain categories in the forms model.