Printing Updated
Added some models, I hope to transfer all the data from one model to another identical one. Changed a lot of code to be able to handle the previous letters for forms and also to be dynamic like the newer code. Eventually want to transition everything to the dynamic labeling for each form.